requirements.txt
googlesearch-python
requests
colorama
urllib3